% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{S2_Intra_BCRM}
\alias{S2_Intra_BCRM}
\title{S2_Intra_BCRM}
\format{
List with 4 elements
}
\source{
Simulation
}
\usage{
S2_Intra_BCRM
}
\description{
S2_Intra_BCRM
}
\keyword{datasets}
