% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{S3_Intra_BOIN}
\alias{S3_Intra_BOIN}
\title{S3_Intra_BOIN}
\format{
List with 4 elements
}
\source{
Simulation
}
\usage{
S3_Intra_BOIN
}
\description{
S3_Intra_BOIN
}
\keyword{datasets}
