% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{S7_Intra_TPT}
\alias{S7_Intra_TPT}
\title{S7_Intra_TPT}
\format{
List with 4 elements
}
\source{
Simulation
}
\usage{
S7_Intra_TPT
}
\description{
S7_Intra_TPT
}
\keyword{datasets}
