% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_trial}
\alias{sim_trial}
\title{Simulates a single trial}
\usage{
sim_trial(
  method,
  tox_rates,
  target,
  cohortsize = 1,
  start = 1,
  bcrm_constrain = NULL,
  bcrm_ff = "power",
  bcrm_method = "exact",
  bcrm_prior = NULL,
  bcrm_stop = NULL,
  boin_ncohort = NULL,
  intra_days = NULL,
  intra_split_tox_rates = TRUE,
  tpt_dose_expansion = TRUE
)
}
\arguments{
\item{method}{Vector of dose-finding designs to be employed. The designs will be applied sequentially with the estimated MTD from the previous design as the starting dose. Possible options are "Intra" for Intra-patient escalation, "TPT" for the 3+3 design, "BCRM" for (Bayesian) Continual Reassessment Method and "BOIN" for Bayesian Optimal Interval Design.}

\item{tox_rates}{Vector of true toxicity rates for each dose level in ascending order}

\item{target}{Target toxicity rate}

\item{cohortsize}{Number of patients in a cohort}

\item{start}{Dose level to be used at the beginning of the trial}

\item{bcrm_constrain}{Compare bcrm::bcrm}

\item{bcrm_ff}{Compare bcrm::bcrm}

\item{bcrm_method}{Compare bcrm::bcrm}

\item{bcrm_prior}{Compare bcrm::bcrm}

\item{bcrm_stop}{Compare bcrm::bcrm}

\item{boin_ncohort}{Compare BOIN::get.oc}

\item{intra_days}{How many single treatments are given}

\item{intra_split_tox_rates}{Boolean value whether toxicity rates are split evenly assuming identical and independent toxicity rates for each treatment. FALSE implies a binary toxicity rate.}

\item{tpt_dose_expansion}{Boolean value whether a dose expansion stage is included if fewer than 6 patients were treated at the estimated MTD}
}
\value{
res Data frame with number of patients, number of DLTs and estimated MTD for each trial

res_dose Data frame with number of patients and number of DLTs for each combination of dose level and trial

true_mtd Dose level which is the closest to target toxicity
}
\description{
Simulates a single trial
}
\examples{

sim_trial(method=c("Intra","TPT"),tox_rates=seq(0.1,0.5,0.1),target=0.3,intra_days=4)
}
