% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{summary_trials}
\alias{summary_trials}
\title{Summarizes the result of a multiple simulated trials}
\usage{
summary_trials(res_list)
}
\arguments{
\item{res_list}{Output from \code{sim_trials}}
}
\description{
Summarizes the result of a multiple simulated trials
}
